+++ src/rosbot_kinematics.h
@@ -37,28 +37,31 @@
 
 #endif
 
+namespace rosbot_kinematics {
+
 extern const RosbotDrive_params_t ROSBOT_PARAMS;
 
 typedef struct RosbotOdometry
 {
-    float wheel_FR_ang_pos;      // radians
-    float wheel_FL_ang_pos;      // radians
-    float wheel_RR_ang_pos;      // radians
-    float wheel_RL_ang_pos;      // radians
-    float wheel_L_ang_pos;       // radians
-    float wheel_R_ang_pos;       // radians
-    float wheel_L_ang_vel = 0;   // radians per second
-    float wheel_R_ang_vel = 0;   // radians per second
-    float robot_angular_pos = 0; // radians
-    float robot_angular_vel = 0; // radians per second
-    float robot_x_pos = 0;       // meters
-    float robot_y_pos = 0;       // meters
-    float robot_x_vel = 0;       // meters per second
-    float robot_y_vel = 0;       // meters per second
+    float wheel_FR_ang_pos;  // radians
+    float wheel_FL_ang_pos;  // radians
+    float wheel_RR_ang_pos;  // radians
+    float wheel_RL_ang_pos;  // radians
+    float wheel_L_ang_pos;   // radians
+    float wheel_R_ang_pos;   // radians
+    float wheel_L_ang_vel;   // radians per second
+    float wheel_R_ang_vel;   // radians per second
+    float robot_angular_pos; // radians
+    float robot_angular_vel; // radians per second
+    float robot_x_pos;       // meters
+    float robot_y_pos;       // meters
+    float robot_x_vel;       // meters per second
+    float robot_y_vel;       // meters per second
 } RosbotOdometry_t;
 
 void setRosbotSpeed(RosbotDrive * drive, float linear, float angular);
 void updateRosbotOdometry(RosbotDrive * drive, RosbotOdometry_t * odom, float dtime);
 void resetRosbotOdometry(RosbotDrive * drive, RosbotOdometry_t * odom);
 
+}
 #endif /* __ROSBOT_KINEMATICS_H__ */