+++ src/rosbot_ros_layer.h
@@ -1,4 +1,0 @@
-#ifndef __ROSBOT_ROS_LAYER__
-#define __ROSBOT_ROS_LAYER__
-
-#endif /* __ROSBOT_ROS_LAYER__ */